requirements.txt
altair==5.5.0
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
category_encoders==2.7.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
dacite==1.9.2
dash==3.0.4
decorator==5.2.1
deprecation==2.1.0
executing==2.2.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
htmlmin==0.1.12
idna==3.10
ImageHash==4.3.1
imbalanced-learn==0.13.0
importlib_metadata==8.7.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kaleido==0.2.1
kiwisolver==1.4.8
lightgbm==4.6.0
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
missingno==0.5.2
multimethod==1.12
narwhals==1.37.1
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.58.1
numpy==1.25.2
orjson==3.10.18
packaging==24.2
pandas==1.5.3
parso==0.8.4
patsy==1.0.1
phik==0.12.4
pillow==10.4.0
platformdirs==4.3.7
plotly==5.24.1
plotly-resampler==0.10.0
pmdarima==2.0.4
prompt_toolkit==3.0.51
protobuf==4.25.7
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pycaret==3.3.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
pyod==2.0.5
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyWavelets==1.8.0
pywin32==310
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.24.0
schemdraw==0.15
scikit-base==0.7.8
scikit-learn==1.4.2
scikit-plot==0.3.7
scipy==1.11.4
seaborn==0.12.2
six==1.17.0
sklearn-compat==0.1.3
sktime==0.26.0
smmap==5.0.2
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.33.0
streamlit-pandas-profiling==0.1.3
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tenacity==8.5.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
tsdownsample==0.1.4.1
typeguard==4.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
visions==0.7.5
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.14
wordcloud==1.9.4
xxhash==3.5.0
ydata-profiling==4.6.4
yellowbrick==1.5
zipp==3.21.0
